+++ client/src/types/myCollections/INotes.tsx
@@ -1,0 +1,9 @@
+interface INotes {
+  content: string
+  label: string
+  labelLanguages?: Array<string>
+  classification: string
+  classificationLanguages?: Array<string>
+}
+
+export default INotes