+++ client/CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Added edit My Collections functionality for classifications ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 - Added edit My Collections functionality for identifiers ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 - Added edit My Collections functionality for web pages ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
+- Added edit My Collections functionality for notes ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 
 ### Changed